health-bot/requirements.txt.txt
fastapi
uvicorn[standard]
langchain
langchain-huggingface
python-dotenv
faiss-cpu
pypdf
langchain-community
langchain-text-splitters
sentence-transformers
pandas
googletrans==4.0.0rc1
langdetect
requests
sqlmodel
aiosqlite
passlib[bcrypt]
python-jose[cryptography]
python-multipart
chromadb
email-validator
pytesseract
Pillow
vosk